Camera control: gp_file_get_data_and_size() takes unsigned long int *
(cherry picked from commit 879b2e6234c2d29ef58194747434e08814305ffe)

+++ src/common/camera_control.c
@@ -901,7 +901,7 @@ void dt_camctl_import(const dt_camctl_t *c, const dt_camera_t *cam, GList *image
       CameraFile* camfile;
       int res = GP_OK;
       const uint8_t* data = NULL;
-      size_t size;
+      unsigned long int size;
 
       gp_file_new(&camfile);
       if ((res = gp_camera_file_get(cam->gpcam, folder, filename, GP_FILE_TYPE_NORMAL, camfile, NULL)) < GP_OK) {
